+++ test/lein_viz/sample.clj
@@ -12,12 +12,12 @@
 
 (defn make-graph
   []
-  {:service [:db :mailer]
-   :db [:datasource]
-   :datasource [:db-host :db-port :database :username :password]
-   :mailer [:smtp-host :smtp-port]})
+  {:data {:service [:db :mailer]
+          :db [:datasource]
+          :datasource [:db-host :db-port :database :username :password]
+          :mailer [:smtp-host :smtp-port]}})
 
 
 (defn make-tree
   []
-  [:foo [10 20] :bar [30 40] :baz [50]])
+  {:data [:foo [10 20] :bar [30 40] :baz [50]]})