+++ CHANGES.md
@@ -5,19 +5,21 @@
 ## 0.2.0 / 2016-September-??
 
 * Payload
-  * [TODO] Data to be under the key `:data`
+  * Data to be under the key `:data`
   * [TODO] Visualization metadata to be under `:layout`
   * [TODO] Authorized seed keys to be under `:seed`
 * Features
   * [TODO] Labeled edges: http://zerosalife.github.io/blog/2014/04/26/clojure-rhizome-labeled-edge-tutorial/
   * [TODO] Zoom into a subgraph by specifying the tip/root at the command line
 * CLI support
+  * Support for choosing selector `["-e" "--selector SELECTOR" "Selector in project.clj"]`
+  * BREAKING CHANGE: Target is now called source `["-s" "--source SOURCE" "Fully qualified defn var name (arity-0)"]`
+  * BREAKING CHANGE: Source may now be placed only under the `:source` key under a selector
+  * Read payload from STDIN when source is `:stdin`
   * [TODO] Flag to show/hide non-existent dependencies
   * [TODO] Flag to show the graph starting from a specified root
   * [TODO] Main action `["-a" "--action ACTION" "Action: view|emit" :default "view" ..]`
   * [TODO] Image type `["-e" "--ext EXTENSION" "Image file extension (file type)" :default "png" ..]`
-  * [TODO] Dependency graph `["-d" "--deps" "Treat graph data as dependency graph (for graph only)"]`
-  * [TODO] Read EDN input from STDIN when target is unspecified or when last arg is `--`
 
 
 ## 0.1.0 / 2016-July-19